Externall app calls

+++ src/test/java/com/mbajdak/reportapp/TestReportFactory.java
@@ -4,8 +4,8 @@ import com.mbajdak.reportapp.domain.Report;
 
 public class TestReportFactory {
     public static Report getTestReport(String characterName, String planetName,
-                                       Long filmId, String filmName,
-                                       Long characterId, Long planetId, Integer reportId) {
+                                       Integer filmId, String filmName,
+                                       Integer characterId, Integer planetId, Integer reportId) {
         Report testReport = new Report();
         testReport.setCharacterName(characterName);
         testReport.setPlanetName(planetName);
